+++ swap/src/bitcoin/redeem.rs
@@ -87,14 +87,20 @@ impl TxRedeem {
             };
 
             // The order in which these are inserted doesn't matter
-            satisfier.insert(A, ::bitcoin::EcdsaSig {
-                sig: sig_a.into(),
-                hash_ty: EcdsaSighashType::All,
-            });
-            satisfier.insert(B, ::bitcoin::EcdsaSig {
-                sig: sig_b.into(),
-                hash_ty: EcdsaSighashType::All,
-            });
+            satisfier.insert(
+                A,
+                ::bitcoin::EcdsaSig {
+                    sig: sig_a.into(),
+                    hash_ty: EcdsaSighashType::All,
+                },
+            );
+            satisfier.insert(
+                B,
+                ::bitcoin::EcdsaSig {
+                    sig: sig_b.into(),
+                    hash_ty: EcdsaSighashType::All,
+                },
+            );
 
             satisfier
         };