+++ swap/src/bitcoin/refund.rs
@@ -70,14 +70,20 @@ impl TxRefund {
             };
 
             // The order in which these are inserted doesn't matter
-            satisfier.insert(A, ::bitcoin::EcdsaSig {
-                sig: sig_a.into(),
-                hash_ty: EcdsaSighashType::All,
-            });
-            satisfier.insert(B, ::bitcoin::EcdsaSig {
-                sig: sig_b.into(),
-                hash_ty: EcdsaSighashType::All,
-            });
+            satisfier.insert(
+                A,
+                ::bitcoin::EcdsaSig {
+                    sig: sig_a.into(),
+                    hash_ty: EcdsaSighashType::All,
+                },
+            );
+            satisfier.insert(
+                B,
+                ::bitcoin::EcdsaSig {
+                    sig: sig_b.into(),
+                    hash_ty: EcdsaSighashType::All,
+                },
+            );
 
             satisfier
         };